Experiment with inlining trace functions, avoiding allocations
This is closer to the final skeleton I imagine; we have a global
variable to enable/disable tracing, and we'd like non-tracing mode to be
fast.

This seems to work correctly (when _TRACING_ENABLED=True) on the
inspection_test hello world, but see concerns.

In this version RTS reports we allocated only 32 bytes less when we
remove the -fplugin=WhatItDo option, which appears to be entirely the
static size associated with _TRACING_ENABLED :: Ptr Bool

The best I could get with a proper NOINLINE unsafePerformIO noop trace
payload was 112 bytes overhead, per instrumented block.